refactor larger-than-n and move utility into the same directory as this exercise
